Update us_exped_botany_week2.Rmd
added some comments to make my choice of objects better or changed the name to make it make more sense, or comments to explain the code a little bit more. I also updated the name from "nmnh" in some places to just "btny" because this isn't the whole nmnh dataset, just botany.